Adds forge 1.18 support, breaking api change and few improvements
