+++ fabric-loader-0.14/src/main/resources/META-INF/services/settingdust.preloadingtricks.LanguageProviderCallback
@@ -1,1 +1,1 @@
-settingdust.preloadingtricks.fabric.FabricLanguageProviderCallback
+settingdust.preloadingtricks.fabric.fabric014.FabricLanguageProviderCallback
