+++ fml-40/src/main/java/settingdust/preloadingtricks/forge/fml40/PreloadingTricksEntrypoint.java
@@ -1,0 +1,6 @@
+package settingdust.preloadingtricks.forge.fml40;
+
+import net.minecraftforge.fml.common.Mod;
+
+@Mod("preloading_tricks")
+public class PreloadingTricksEntrypoint {}
