+++ fml-40/src/main/resources/META-INF/services/settingdust.preloadingtricks.LanguageProviderCallback
@@ -1,0 +1,1 @@
+settingdust.preloadingtricks.forge.fml40.ForgeLanguageProviderCallback
