--- fml-45/src/main/java/settingdust/preloadingtricks/forge/PreloadingTricksEntrypoint.java
@@ -1,6 +1,0 @@
-package settingdust.preloadingtricks.forge;
-
-import net.minecraftforge.fml.common.Mod;
-
-@Mod("preloading_tricks")
-public class PreloadingTricksEntrypoint {}
