+++ fml-47/build.gradle.kts
@@ -1,12 +1,22 @@
-import net.fabricmc.loom.task.AbstractRunTask
-
 plugins {
     alias(libs.plugins.architectury)
     alias(libs.plugins.architectury.loom)
 }
 
 architectury {
+    platformSetupLoomIde()
     forge()
+}
+
+val mod_id: String by rootProject
+
+loom {
+    mods {
+        named("main") {
+            modSourceSets.empty()
+            modFiles.setFrom(tasks.jar)
+        }
+    }
 }
 
 dependencies {
@@ -19,12 +29,19 @@ dependencies {
     implementation(project(":preloading-callbacks")) {
         isTransitive = false
     }
+
+    implementation(project(":forge-language-provider"))
+    include(project(":forge-language-provider"))
 }
 
 tasks {
+    classes {
+        finalizedBy(jar)
+    }
+
     jar {
         manifest.attributes(
-            "FMLModType" to "LANGPROVIDER"
+            "FMLModType" to "LIBRARY"
         )
     }
 }
