+++ fml-47/gradle.properties
@@ -1,0 +1,1 @@
+loom.platform=forge
