+++ fml-47/src/main/java/settingdust/preloadingtricks/forge/fml47/PreloadingTricksEntrypoint.java
@@ -1,0 +1,6 @@
+package settingdust.preloadingtricks.forge.fml47;
+
+import net.minecraftforge.fml.common.Mod;
+
+@Mod("preloading_tricks")
+public class PreloadingTricksEntrypoint {}
