+++ fml-47/src/main/resources/pack.mcmeta
@@ -1,0 +1,6 @@
+{
+  "pack": {
+    "description": "Preloading Tricks Dummy Resources",
+    "pack_format": 4
+  }
+}
