+++ gradle.properties
@@ -2,7 +2,7 @@ kotlin.code.style=official
 # Gradle Properties
 org.gradle.jvmargs=-Xmx4G
 
-version=0.0.12
+version=0.1.0
 group=settingdust.preloadingtricks
 archives_name=preloading-tricks
