+++ preloading-callbacks/src/main/java/settingdust/preloadingtricks/util/ServiceLoaderUtil.java
@@ -7,7 +7,7 @@ import java.util.ServiceLoader;
 public class ServiceLoaderUtil {
 
     public static <T> void loadServices(Class<T> clazz, ServiceLoader<T> serviceLoader, Logger logger) {
-        final var prefix = String.format("[%s]", logger.getName());
+        final var prefix = String.format("[%s] ", logger.getName());
         final var iterator = serviceLoader.stream().iterator();
         var hasNext = false;
         var empty = false;
@@ -26,16 +26,24 @@ public class ServiceLoaderUtil {
             final var provider = iterator.next();
 
             String providerName = provider.type().getName();
+            boolean isInternal = providerName.startsWith("settingdust.preloadingtricks.");
 
-            logger.info(prefix + "Loading " + providerName);
+            if (isInternal) {
+                logger.debug(prefix + "Loading " + providerName);
+            } else {
+                logger.info(prefix + "Loading " + providerName);
+            }
 
             try {
                 provider.get();
             } catch (Throwable t) {
-                if (!providerName.startsWith("settingdust.preloadingtricks.")) {
-                    logger.error(prefix + "Loading " + providerName + " failed");
+                if (isInternal) {
+                    logger.debug(prefix + "Loading " + providerName + " failed", t);
+                } else {
+                    String error = prefix + "Loading " + providerName + " failed";
+                    logger.error(error, t);
+                    throw new RuntimeException(error, t);
                 }
-                logger.debug(prefix + "Loading " + providerName + " failed", t);
             }
 
             try {
