--- quilt-loader-0.20/src/main/resources/META-INF/services/settingdust.preloadingtricks.LanguageProviderCallback
@@ -1,1 +1,0 @@
-settingdust.preloadingtricks.quilt.QuiltLanguageProviderCallback
