+++ quilt-loader-0.21/src/main/resources/META-INF/services/settingdust.preloadingtricks.LanguageProviderCallback
@@ -1,0 +1,1 @@
+settingdust.preloadingtricks.quilt.quilt021.QuiltLanguageProviderCallback
