+++ settings.gradle.kts
@@ -20,7 +20,6 @@ pluginManagement {
             url = uri("https://maven.minecraftforge.net/")
         }
         gradlePluginPortal()
-        mavenLocal()
     }
 }
 
@@ -33,8 +32,9 @@ rootProject.name = "preloading-tricks"
 include("preloading-callbacks")
 
 include("fabric-loader-0.14")
-include("quilt-loader-0.20")
+include("quilt-loader-0.21")
 include("fabric-like-language-adapter")
 
-include("fml-45")
+include("fml-47")
+include("fml-40")
 include("forge-language-provider")
